Switch from using timeouts to an explicit ReadMode.
There were many questions about timeouts and the exact reading mode
configuration of a `SerialPort`. Rather than continue to use a `timeout`
setting, switch to using a more understandable `ReadMode` enum.
Additionally make it a required argument for opening a port to ensure
users understand the implications of this choice.